src/leetcode/FindMinimumInRotatedSortedArray_153.java
/*
 * Copyright (c) 2024 THGEEK.COM. All rights reserved.
 */

package leetcode;

/**
 * LeetCode 153
 * Difficulty: Medium
 *
 * @author Tao Hong
 * @version 1.0
 * @since 2024/08/08 11:45
 */
public class FindMinimumInRotatedSortedArray_153 {
    public int findMin(int[] nums) {
        int n = nums.length;
        int left = 0;
        int right = n - 1;
        while (left < right) {
            if (nums[left] < nums[right]) {
                return nums[left];
            }
            int mid = left + (right - left) / 2;
            if (nums[mid] >= nums[left]) {
                left = mid + 1;
            } else {
                right = mid;
            }
        }
        return nums[left];
    }
}
